major bug fix in global connection
When I made the connection global, I introduced a major bug: if `vega`
isn't running when the app starts up, the connection is never
reallocated. Fix this with a combination of changes both here and in
the myria-python submodule, so that we don't require the connection to
be live when the MyriaConnection object is allocated.