editor: fall back to relational plan visualizer
We cannot yet visualize the physical plans that include Sequence or
DoWhile operators using the new hotness. For these queries, we can only
show the relational plan using the old coldness.

Catch errors visualizing physical plans in the general case, and when
they occur hide the physical plan and show the relational plan.

Vice versa -- when parsing a new query and hence visualizing a new plan,
if the visualizer succeeds then hide the relational plan and show the
physical plan.